Separate sync and async code
