+++ Gruntfile.js
@@ -23,6 +23,6 @@ module.exports = function(grunt) {
     });
     grunt.loadNpmTasks('grunt-contrib-jshint');
     grunt.loadNpmTasks('grunt-mocha-test');
-    grunt.registerTask('default', ['jshint', 'mochaTest']);
+    grunt.registerTask('default', ['mochaTest']);
     grunt.registerTask('api', ['mochaTest']);
 };
